requirements.txt
requests
python-dotenv
logzero
boto3
pyYAML
